Refactor ak_ingest: Split ingest.py into lib.py, simplify main entry
